Put handlers into it's own module and cleanup
